refactor: merge add/accept resources
