+++ tests/package.json
@@ -38,8 +38,7 @@
     "testGetOwnedNfts": "mocha --timeout 9999999 -r ts-node/register './src/getOwnedNfts.test.ts'",
     "testSetNftProperty": "mocha --timeout 9999999 -r ts-node/register './src/setNftProperty.test.ts'",
     "testSetCollectionProperty": "mocha --timeout 9999999 -r ts-node/register './src/setCollectionProperty.test.ts'",
-    "testAddNftResource": "mocha --timeout 9999999 -r ts-node/register './src/addNftResource.test.ts'",
-    "testAcceptNftResource": "mocha --timeout 9999999 -r ts-node/register './src/acceptNftResource.test.ts'",
+    "testAddResource": "mocha --timeout 9999999 -r ts-node/register './src/addResource.test.ts'",
     "testRemoveResource": "mocha --timeout 9999999 -r ts-node/register './src/removeResource.test.ts'",
     "testSetResourcePriorities": "mocha --timeout 9999999 -r ts-node/register './src/setResourcePriorities.test.ts'",
     "testSetEquippableList": "mocha --timeout 9999999 -r ts-node/register './src/setEquippableList.test.ts'",
